refactoring: выделил абстрактный класс для view
