+++ .eslintrc.yml
@@ -8,3 +8,8 @@ env:
   commonjs: true
 
 extends: 'airbnb-base'
+
+rules:
+  class-methods-use-this: [error, {
+    exceptMethods: ['getMarkup']
+  }]
