+++ src/js/view/header.js
@@ -1,11 +1,6 @@
-import createElement from '../helpers/createElement';
+import Abstract from './abstract';
 
-export default class Header {
-  constructor() {
-    this.element = null;
-    this.markup = null;
-  }
-
+export default class Header extends Abstract {
   getMarkup() {
     this.markup = (
       `<div class="container">
@@ -35,17 +30,5 @@ export default class Header {
     );
 
     return this.markup;
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getMarkup());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
