+++ src/js/view/locations.js
@@ -1,11 +1,6 @@
-import createElement from '../helpers/createElement';
+import Abstract from './abstract';
 
-export default class About {
-  constructor() {
-    this.element = null;
-    this.markup = null;
-  }
-
+export default class About extends Abstract {
   getMarkup() {
     this.markup = (
       `<section class="locations">
@@ -138,17 +133,5 @@ export default class About {
     );
 
     return this.markup;
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getMarkup());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
